*improving of SSSE3 and AVX2 optimizations of function GaussianBlur3x3.

+++ src/Simd/SimdBaseHistogram.cpp
@@ -119,7 +119,7 @@ namespace Simd
                 if (sum == 0 && histogram[i] != 0)
                 {
                     minCount = histogram[i]; 
-                    minColor = i;
+                    minColor = (uint32_t)i;
                 }
                 sum += histogram[i];
                 integral[i] = sum;
